Code is seperated into different files
