+++ src/client.cpp
@@ -1,0 +1,6 @@
+#include "client.h"
+
+#include <string>
+
+
+Client::Client(std::string name ) : clientName(name) {}
