+++ src/console_ui.cpp
@@ -1,0 +1,6 @@
+#include "console_ui.h"
+
+#include <stdlib.h>
+#include <iostream>
+#include <string>
+
