+++ src/utils.cpp
@@ -1,0 +1,14 @@
+
+#include "utils.h"
+
+enum class Currency {
+    USD,
+    EUR,
+    GBP
+};
+
+struct Transaction {
+    double amount;
+    Currency from;
+    Currency to;
+};
